Process of cutting off golog's mode_head into mode_head.py (Stable but not working as intended)
